+++ src/external.hpp
@@ -260,7 +260,7 @@ struct External {
 
   void reset_assumptions ();
 
-  // similarily to 'failed', 'conclude' needs to know about failing
+  // similarly to 'failed', 'conclude' needs to know about failing
   // assumptions and therefore needs to be reset when leaving the
   // 'UNSATISFIED' state.
   //